feat: load favicon if configured via params.favicon
This will display a favicon for the website. Favicons are cute little
icons that will show up on tabs in your browser, e.g. GitHub's mascot.
They can easily be generated by using online tools, e.g.

https://realfavicongenerator.net/

The resulting package only has to be downloaded and extracted into the
'static' directory of a Hugo website that is using this theme.

The path can then be set as follows in the config.yml:

```yaml
params:
  favicon: favicon.ico
```

This fixes #3.