feat: add params.navbar_style to config to select a style for the navbar
This is a new feature in readable.css 1.1.0. The new version of
readable.css defaults to a navbar without animations, but to stay
compatible with older releases of this theme, we default to "classy"
instead which is the animated navbar.

The new navbar styles also look better without wrapping the elements
inside a `<span>`, so we make that conditional and only apply it when
the "classy" style is chosen.