CI: Add GitHub workflow to check for readable.css updates
This adds a workflow and accompanying script to update readable.css. The
script gets the latest tag version from the readable.css repository on
codeberg.org and compares it to the currently used version. If the
versions don't match up, we download readable.css and readable.min.css
from Codeberg and update layouts/partials/head.html.

I initially didn't bother to check the current version, because
ultimately we only do 3 requests against Codeberg every time the
workflow executes. However, I'm expecting this workflow to run once a
day currently. If someone were to increase the frequency significantly,
this might become problematic for Codeberg. It's easy enough to
optimize, so we should do it.

The workflow uses 'peter-evans/create-pull-request@v5', an action that
allows us to create GitHub pull requests on the repo.

The action seems really thought through to me. I checked that an
existing pull request is not duplicated. In case the 'main' branch
changes in between, it will simply force push changes to the existing
pull request. This also means that we won't get flooded with pull
requests for the same version update.

If no changes are found, then the action will not open up a pull request
at all.

The repository's settings need to be adjusted to allow workflows to
create pull requests. This can be done in GitHub, by opening up the
repository's 'Settings', clicking 'Actions' and 'General' and updating
'Workflow permissions' to allow 'Read and write permissions' and 'Allow
GitHub Actions to create and approve pull requests'.

Other than that, there's nothing to do to enable this workflow. Once
this commit is merged, the workflow will start it's scheduled execution.

The only thing I couldn't seem to get working is a notification for the
newly created pull request. I'm not sure if I was simply unable to get
it to work with a fork, or what's the reason behind this. I would have
expected an email at least. However, I think this can be optimized in
a future commit.