CI: Add GitHub workflow to check for readable.css updates (#15)
This adds a workflow and accompanying script to update readable.css. The
script gets the latest tag version from the readable.css repository on
codeberg.org and compares it to the currently used version. If the
versions don't match up, we download readable.css and readable.min.css
from Codeberg and update layouts/partials/head.html.